+++ src/eradius_client.erl
@@ -14,7 +14,8 @@
 -export([start_link/0, send_request/2, send_request/3, send_remote_request/3, send_remote_request/4]).
 %% internal
 -export([reconfigure/0, send_remote_request_loop/8, find_suitable_peer/1,
-         restore_upstream_server/1]).
+         restore_upstream_server/1, store_radius_server_from_pool/3,
+         init_server_status_metrics/0]).
 
 -behaviour(gen_server).
 -export([init/1, handle_call/3, handle_cast/2, handle_info/2, terminate/2, code_change/3]).
@@ -67,26 +68,47 @@ send_request({Host, Port, Secret}, Request, Options)
     send_request({erlang:binary_to_list(Host), Port, Secret}, Request, Options);
 send_request({Host, Port, Secret}, Request, Options) 
   when ?GOOD_CMD(Request) andalso is_list(Host) ->
-    case inet:gethostbyname(Host) of
-        {ok, #hostent{h_addrtype = inet, h_addr_list = [IP]}} -> 
-            send_request({IP, Port, Secret}, Request, Options);
-        {ok, #hostent{h_addrtype = inet, h_addr_list = [_ | _] = IPs}} -> 
-            Index = rand:uniform(length(IPs)),
-            IP = lists:nth(Index, IPs),
-            send_request({IP, Port, Secret}, Request, Options);
-        _ -> error(badarg)
-    end;
+    IP = get_ip(Host),
+    send_request({IP, Port, Secret}, Request, Options);
 send_request({IP, Port, Secret}, Request, Options) when ?GOOD_CMD(Request) andalso is_tuple(IP) ->
-    TS1 = eradius_lib:timestamp(milli_seconds),
+    TS1 = erlang:monotonic_time(),
     ServerName = proplists:get_value(server_name, Options, undefined),
     MetricsInfo = make_metrics_info(Options, {IP, Port}),
-    update_client_requests(MetricsInfo),
     Retries = proplists:get_value(retries, Options, ?DEFAULT_RETRIES),
     Timeout = proplists:get_value(timeout, Options, ?DEFAULT_TIMEOUT),
-    Peer = {ServerName, {IP, Port}},
-    {Socket, ReqId} = gen_server:call(?SERVER, {wanna_send, Peer, MetricsInfo}),
-    Response = send_request_loop(Socket, ReqId, Peer, Request#radius_request{reqid = ReqId, secret = Secret}, Retries, Timeout, MetricsInfo),
-    proceed_response(Request, Response, Peer, TS1, MetricsInfo, Options);
+    SendReqFn = fun () ->
+        Peer = {ServerName, {IP, Port}},
+        update_client_requests(MetricsInfo),
+        {Socket, ReqId} = gen_server:call(?SERVER, {wanna_send, Peer, MetricsInfo}),
+        Response = send_request_loop(Socket, ReqId, Peer,
+                                     Request#radius_request{reqid = ReqId, secret = Secret},
+                                     Retries, Timeout, MetricsInfo),
+        proceed_response(Request, Response, Peer, TS1, MetricsInfo, Options)
+    end,
+    % If we have other RADIUS upstream servers check current one,
+    % maybe it is already marked as inactive and try to find another
+    % one
+    case proplists:get_value(failover, Options, []) of
+        [] ->
+            SendReqFn();
+        UpstreamServers ->
+            case find_suitable_peer([{IP, Port, Secret} | UpstreamServers]) of
+                [] ->
+                    no_active_servers;
+                {{IP, Port, Secret}, _NewPool} ->
+                    SendReqFn();
+                {NewPeer, []} ->
+                    % Special case, we don't have servers in the pool anymore, but we need
+                    % to preserve `failover` option to mark current server as inactive if
+                    % it will fail
+                    NewOptions = lists:keyreplace(failover, 1, Options, {failover, undefined}),
+                    send_request(NewPeer, Request, NewOptions);
+                {NewPeer, NewPool} ->
+                    % current server is not in list of active servers, so use another one
+                    NewOptions = lists:keyreplace(failover, 1, Options, {failover, NewPool}),
+                    send_request(NewPeer, Request, NewOptions)
+            end
+    end;
 send_request({_IP, _Port, _Secret}, _Request, _Options) ->
     error(badarg).
 
@@ -100,7 +122,7 @@ send_remote_request(Node, NAS, Request) ->
 %   The request will not be sent again if the remote node is unreachable.
 -spec send_remote_request(node(), nas_address(), #radius_request{}, options()) -> {ok, binary()} | {error, 'timeout' | 'node_down' | 'socket_down'}.
 send_remote_request(Node, {IP, Port, Secret}, Request, Options) when ?GOOD_CMD(Request) ->
-    TS1 = eradius_lib:timestamp(milli_seconds),
+    TS1 = erlang:monotonic_time(),
     ServerName = proplists:get_value(server_name, Options, undefined),
     MetricsInfo = make_metrics_info(Options, {IP, Port}),
     update_client_requests(MetricsInfo),
@@ -138,7 +160,7 @@ restore_upstream_server({ServerIP, Port, Retries, InitialRetries}) ->
     ets:insert(?MODULE, {{ServerIP, Port}, Retries, InitialRetries}).
 
 proceed_response(Request, {ok, Response, Secret, Authenticator}, _Peer = {_ServerName, {ServerIP, Port}}, TS1, MetricsInfo, Options) ->
-    update_client_request(Request#radius_request.cmd, MetricsInfo, eradius_lib:timestamp(milli_seconds) - TS1, Request),
+    update_client_request(Request#radius_request.cmd, MetricsInfo, erlang:monotonic_time() - TS1, Request),
     update_client_responses(MetricsInfo),
     case eradius_lib:decode_request(Response, Secret, Authenticator) of
         {bad_pdu, "Message-Authenticator Attribute is invalid" = Reason} ->
@@ -158,16 +180,18 @@ proceed_response(Request, {ok, Response, Secret, Authenticator}, _Peer = {_Serve
             ?LOG(error, "~s INF: Noreply for request ~p. Could not decode the request, reason: ~s", [printable_peer(ServerIP, Port), Request, Reason]),
             maybe_failover(Request, noreply, {ServerIP, Port}, Options);
         Decoded ->
+            update_server_status_metric(ServerIP, Port, true, Options),
             update_client_response(Decoded#radius_request.cmd, MetricsInfo, Request),
             {ok, Response, Authenticator}
     end;
 
 proceed_response(Request, Response, {_ServerName, {ServerIP, Port}}, TS1, MetricsInfo, Options) ->
     update_client_responses(MetricsInfo),
-    update_client_request(Request#radius_request.cmd, MetricsInfo, eradius_lib:timestamp(milli_seconds) - TS1, Request),
+    update_client_request(Request#radius_request.cmd, MetricsInfo, erlang:monotonic_time() - TS1, Request),
     maybe_failover(Request, Response, {ServerIP, Port}, Options).
 
 maybe_failover(Request, Response, {ServerIP, Port}, Options) ->
+    update_server_status_metric(ServerIP, Port, false, Options),
     case proplists:get_value(failover, Options, []) of
         [] ->
             Response;
@@ -194,13 +218,7 @@ handle_failed_request(Request, {ServerIP, Port} = _FailedServer, UpstreamServers
                     ets:update_counter(?MODULE, {ServerIP, Port}, -FailedTries)
             end;
         [] ->
-            % That could be an upstream server defined as default proxy relay or just
-            % just RADIUS server that was used by default via send_request/3,
-            % so initially such servers is not in the pool. So just skip it for now,
-            % and put it into ets later
-            Timeout = application:get_env(eradius, unreachable_timeout, 60),
-            timer:apply_after(Timeout * 1000, ?MODULE, restore_upstream_server,
-                              [{ServerIP, Port, ?DEFAULT_RETRIES, ?DEFAULT_RETRIES}])
+            ok
     end,
     case find_suitable_peer(UpstreamServers) of
         [] ->
@@ -223,7 +241,8 @@ send_request_loop(Socket, ReqId, Peer, Request, Retries, Timeout, MetricsInfo) -
     send_request_loop(Socket, SMon, Peer, ReqId, Authenticator, EncRequest, Timeout, Retries, MetricsInfo, Request#radius_request.secret, Request).
 
 send_request_loop(_Socket, SMon, _Peer, _ReqId, _Authenticator, _EncRequest, Timeout, 0, MetricsInfo, _Secret, Request) ->
-    update_client_request(timeout, MetricsInfo, Timeout, Request),
+    TS = erlang:convert_time_unit(Timeout, millisecond, native),
+    update_client_request(timeout, MetricsInfo, TS, Request),
     erlang:demonitor(SMon, [flush]),
     {error, timeout};
 send_request_loop(Socket, SMon, Peer = {_ServerName, {IP, Port}}, ReqId, Authenticator, EncRequest, Timeout, RetryN, MetricsInfo, Secret, Request) ->
@@ -239,7 +258,8 @@ send_request_loop(Socket, SMon, Peer = {_ServerName, {IP, Port}}, ReqId, Authent
             {error, Error}
     after
         Timeout ->
-            update_client_request(retransmission, MetricsInfo, Timeout, Request),
+            TS = erlang:convert_time_unit(Timeout, millisecond, native),
+            update_client_request(retransmission, MetricsInfo, TS, Request),
             send_request_loop(Socket, SMon, Peer, ReqId, Authenticator, EncRequest, Timeout, RetryN - 1, MetricsInfo, Secret, Request)
     end.
 
@@ -385,7 +405,10 @@ configure(State) ->
 %% private
 prepare_pools() ->
     ets:new(?MODULE, [ordered_set, public, named_table, {keypos, 1}, {write_concurrency,true}]),
-    lists:foreach(fun({_PoolName, Servers}) -> prepare_pool(Servers) end, application:get_env(eradius, servers_pool, [])).
+    lists:foreach(fun({_PoolName, Servers}) -> prepare_pool(Servers) end, application:get_env(eradius, servers_pool, [])),
+    lists:foreach(fun(Server) -> store_upstream_servers(Server) end, application:get_env(eradius, servers, [])),
+    init_server_status_metrics().
+
 prepare_pool([]) -> ok;
 prepare_pool([{Addr, Port, _, Opts} | Servers]) ->
     Retries = proplists:get_value(retries, Opts, ?DEFAULT_RETRIES),
@@ -395,11 +418,36 @@ prepare_pool([{Addr, Port, _} | Servers]) ->
     store_radius_server_from_pool(Addr, Port, ?DEFAULT_RETRIES),
     prepare_pool(Servers).
 
+store_upstream_servers({Server, _}) ->
+    store_upstream_servers(Server);
+store_upstream_servers({Server, _, _}) ->
+    store_upstream_servers(Server);
+store_upstream_servers(Server) ->
+    HandlerDefinitions = application:get_env(eradius, Server, []),
+    UpdatePoolFn = fun (HandlerOpts) ->
+        {DefaultRoute, Routes, Retries} = eradius_proxy:get_routes_info(HandlerOpts),
+        eradius_proxy:put_default_route_to_pool(DefaultRoute, Retries),
+        eradius_proxy:put_routes_to_pool(Routes, Retries)
+    end,
+    lists:foreach(fun (HandlerDefinition) ->
+        case HandlerDefinition of
+            {{_, []}, _} ->             ok;
+            {{_, _, []}, _} ->          ok;
+            {{_, HandlerOpts}, _} ->    UpdatePoolFn(HandlerOpts);
+            {{_, _, HandlerOpts}, _} -> UpdatePoolFn(HandlerOpts);
+            _HandlerDefinition ->       ok
+        end
+    end,
+    HandlerDefinitions).
+
 %% private
 store_radius_server_from_pool(Addr, Port, Retries) when is_tuple(Addr) and is_integer(Port) and is_integer(Retries) ->
     ets:insert(?MODULE, {{Addr, Port}, Retries, Retries});
-store_radius_server_from_pool(Addr, _, _) ->
-    ?LOG(error, "bad IP address specified in RADIUS servers pool configuration ~p", [Addr]),
+store_radius_server_from_pool(Addr, Port, Retries) when is_list(Addr) and is_integer(Port) and is_integer(Retries) ->
+    IP = get_ip(Addr),
+    ets:insert(?MODULE, {{IP, Port}, Retries, Retries});
+store_radius_server_from_pool(Addr, Port, Retries) ->
+    ?LOG(error, "bad RADIUS upstream server specified in RADIUS servers pool configuration ~p", [{Addr, Port, Retries}]),
     error(badarg).
 
 configure_address(State = #state{socket_ip = OAdd, sockets = Sockts}, NPorts, NAdd) ->
@@ -493,6 +541,19 @@ parse_ip(T = {_, _, _, _}) ->
 parse_ip(T = {_, _, _, _, _, _}) ->
     {ok, T}.
 
+init_server_status_metrics() ->
+    case application:get_env(eradius, server_status_metrics_enabled, false) of
+        false ->
+            ok;
+        true ->
+            % That will be called at eradius startup and we must be sure that prometheus
+            % application already started if server status metrics supposed to be used
+            application:ensure_all_started(prometheus),
+            ets:foldl(fun ({{Addr, Port}, _, _}, _Acc) ->
+                eradius_counter:set_boolean_metric(server_status, [Addr, Port], false)
+            end, [], ?MODULE)
+    end.
+
 make_metrics_info(Options, {ServerIP, ServerPort}) ->
     ServerName = proplists:get_value(server_name, Options, undefined),
     ClientName = proplists:get_value(client_name, Options, undefined),
@@ -515,6 +576,29 @@ inc_responses_counter_accounting(MetricsInfo, #radius_request{attrs = Attrs}) ->
     ok;
 inc_responses_counter_accounting(_, _) ->
     ok.
+
+update_server_status_metric(IP, Port, false, _Options) ->
+    eradius_counter:set_boolean_metric(server_status, [IP, Port], false);
+update_server_status_metric(IP, Port, true, Options) ->
+    UpstreamServers = proplists:get_value(failover, Options, []),
+    % set all servesr from pool as inactive
+    if is_list(UpstreamServers) ->
+        lists:foreach(fun (Server) ->
+            case Server of
+                {ServerIP, ServerPort, _} ->
+                    eradius_counter:set_boolean_metric(server_status, [ServerIP, ServerPort], false);
+                {ServerIP, ServerPort, _, _} ->
+                    eradius_counter:set_boolean_metric(server_status, [ServerIP, ServerPort], false);
+                _ ->
+                    ok
+            end
+
+        end, UpstreamServers);
+    true ->
+        ok
+    end,
+    % set current service as active
+    eradius_counter:set_boolean_metric(server_status, [IP, Port], true).
 
 %% check if we can use persistent_term for config
 %% persistent term was added in OTP 21.2 but we can't
@@ -576,6 +660,8 @@ client_response_counter_account_match_spec_compile() ->
 
 -endif.
 
+find_suitable_peer(undefined) ->
+    [];
 find_suitable_peer([]) ->
     [];
 find_suitable_peer([{IP, Port, Secret} | Pool]) ->
@@ -587,3 +673,13 @@ find_suitable_peer([{IP, Port, Secret} | Pool]) ->
     end;
 find_suitable_peer([{IP, Port, Secret, _Opts} | Pool]) ->
     find_suitable_peer([{IP, Port, Secret} | Pool]).
+
+get_ip(Host) ->
+    case inet:gethostbyname(Host) of
+        {ok, #hostent{h_addrtype = inet, h_addr_list = [IP]}} ->
+            IP;
+        {ok, #hostent{h_addrtype = inet, h_addr_list = [_ | _] = IPs}} ->
+            Index = rand:uniform(length(IPs)),
+            lists:nth(Index, IPs);
+        _ -> error(badarg)
+    end.